% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CG.R
\name{CG}
\alias{CG}
\title{Truncated conjugate gradient using Hessian-vector products}
\usage{
CG(
  b,
  Hq,
  x = 0 * b,
  max.it = length(b),
  e = 1e-10,
  Minv = NULL,
  stop_if_nonPD = TRUE,
  silent = TRUE
)
}
\arguments{
\item{b}{vector to solve for}

\item{Hq}{efficient Hessian-vector product function, e.g., \link{make_Hq}}

\item{x}{initial guess for solution}

\item{max.it}{maximum iterations (often should be less than \code{length(b)})}

\item{e}{error criterion}

\item{Minv}{preconditioner matrix (uses Identity by default)}

\item{stop_if_nonPD}{whether to stop CG if recusion is not positive definite}

\item{silent}{Be silent or print progress?}
}
\description{
Approximate the soution to a linear system \eqn{Hq = b} using truncated conjugate gradient,
without forming the Hessian matrix directly and instead using Hessian-vector products
e.g., using forwards-on-reverse autodiff
Adapted from \code{mcmcsae::CG} under GPL-3 licence to use Hessian-vector products
}
\details{
Note that \code{CG} can only be taped when \code{stop_if_nonPD = FALSE} and
\code{e = 0}, and using a fixed number of steps, e.g., \code{max.it = 30}
}
\examples{
library(Matrix)

# Settings
n = 10^4
rho = 0.99

# Simulate AR1 process approaching random walk (i.e., ill-conditioned inner problem)
P = bandSparse( n = n, k = c(-1), diagonals = list(rep(1,n)) )
Q = (Diagonal(n) - rho * t(P)) \%*\% (Diagonal(n) - rho * P)
x = RTMB:::rgmrf0( n= 1, Q = Q )[,1]
y = x + 0.1 * rnorm(n)
which_seen = sample( seq_len(n), size = n/10, replace = FALSE)
y[-which_seen] = NA

nll = function(p){
  -dgmrf(p$x, Q = Q, log = TRUE) - sum(dnorm(y, p$x, sd = 0.1, log=TRUE), na.rm=TRUE)
}
parlist = list( x=rnorm(n) )

tape = MakeTape(nll, parlist)
gr = tape$jacfun()
Hq = make_Hq( tape, x = unlist(parlist) )
H = gr$jacfun(sparse = TRUE)

# Test unregularized solution
b = gr(unlist(parlist))[1,]
Hess = H(unlist(parlist))
x = solve( Hess, b)
out = CG(
  b = b,
  Hq = Hq
)
plot( x, out$x )

# Test trust-region regularization
Hess2 = Hess + 2 * Diagonal(n=n)
x2 = solve( Hess2, b)
out = CG(
  b = b,
  Hq = \(b,...) Hq(b,...) + 2*b
)
plot( x2, out$x )

}
