% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmrf.R
\name{dgmrf_icpl}
\alias{dgmrf_icpl}
\title{Approximate the log-density of a GMRF using incomplete Cholesky}
\usage{
dgmrf_icpl(x, mu = 0, Q, k, v_m, log = FALSE)
}
\description{
Approximate the log-density of a GMRF using incomplete Cholesky
}
