% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmrf.R
\name{dgmrf_lanczos}
\alias{dgmrf_lanczos}
\title{Approximate the log-density of a GMRF using Lanczos}
\usage{
dgmrf_lanczos(
  x,
  Q,
  k,
  mu = rep(0, length(x)),
  m = 3,
  log = FALSE,
  seed = NULL,
  orthogonalize = TRUE
)
}
\arguments{
\item{x}{parameter vector used when calculating the Hessian matrix}

\item{Q}{Sparse precision matrix}

\item{k}{dimension for Kyrlov subspace}

\item{mu}{Mean parameter vector}

\item{m}{number of probe-vectors to use for approximating average and standard
deviation of log-determinant}

\item{log}{Logical; Return log density?}

\item{seed}{if not NULL, then sets the seed.  This is helfpul given that
the Hutchinson probe vectors are randomly sampled, and comparisons have
lower variance using a fixed seed.}

\item{orthogonalize}{Whether to do two-pass Gram-Schmidt re-normalization
(much slower)}
}
\description{
Calculate log-density of a GMRF using Lanczos recursion
}
\examples{
library(RTMB)
library(Matrix)
library(lanczosRTMB)

# Create a precision for a 1D AR1 process
nx = 1e5
P = 0.5 * bandSparse( n = nx, k = c(-1,1) )
Q = (Diagonal(nx) - 0.5*t(P) ) \%*\% (Diagonal(nx) - 0.5*P)
x = RTMB:::rgmrf0(n = 1, Q )[,1]

# Exact log-density using RTMB
dgmrf(x, Q = Q, log = TRUE)

# log-density using stochastic Lanczos quadrature
dgmrf_lanczos(x, Q = Q, k = 20, log = TRUE)

# Log-density using incomplete Cholesky
# dgmrf_icpl(x, Q = Q, log = TRUE)

}
