% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CG.R
\name{estimate_diag_hutchinson}
\alias{estimate_diag_hutchinson}
\title{Estimate diag(H) using Hutchinson probes}
\usage{
estimate_diag_hutchinson(
  Hq,
  n_probes = 30,
  x = attr(Hq, "env")$x0,
  n = length(attr(Hq, "env")$which_random)
)
}
\description{
Estimate diag(H) using Hutchinson probes
}
