% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lanczos.R
\name{lanczos}
\alias{lanczos}
\title{Calculate Lanczos approximation}
\usage{
lanczos(
  Hq,
  q,
  k,
  x = attr(Hq, "env")$x0,
  orthogonalize = FALSE,
  V = matrix(1, nrow = length(x), ncol = 0),
  tol = 1e-12
)
}
\arguments{
\item{Hq}{function that calculates the product \code{H \%*\% q} given probe \code{q} and parameters \code{x}}

\item{q}{initial vector used for defining a Kyrlov subspace}

\item{k}{dimension for Kyrlov subspace}

\item{x}{parameter vector used when calculating the Hessian matrix}

\item{orthogonalize}{Whether to do two-pass Gram-Schmidt re-normalization
(much slower)}

\item{V}{deflation matrix, where we transform probes \eqn{q' = (I-V V^T) q}
and the Hessian by \eqn{(I-V V^T) H (I-V V^T)} to eliminate axes in \eqn{V}}

\item{tol}{numerical tolerance for stopping algorithm given that no more terms are identifiable}
}
\description{
Calculate Q, alpha and beta for Lanzos approximation
}
\examples{
H = diag(exp(rnorm(5)))
q = rep(1,5)
Hq = function(q, x) (H \%*\% q)[,1]

L = lanczos(Hq, x = NULL, q, k = 5, ortho = TRUE)
T = tridiag(L$alpha, L$beta)

# Should match H if and only if L$m = nrow(H)
range(L$Q \%*\% T \%*\% t(L$Q) - H)

}
