% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lanczos_gmrf.R
\name{lanczos_dgmrf}
\alias{lanczos_dgmrf}
\title{Approximate the log-density of a GMRF using Lanczos}
\usage{
lanczos_dgmrf(
  x,
  Q,
  k,
  mu = rep(0, length(x)),
  m = 3,
  log = FALSE,
  seed = NULL,
  orthogonalize = TRUE
)
}
\arguments{
\item{x}{parameter vector used when calculating the Hessian matrix}

\item{Q}{Sparse precision matrix}

\item{k}{dimension for Kyrlov subspace}

\item{mu}{Mean parameter vector}

\item{m}{number of probe-vectors to use for approximating average and standard
deviation of log-determinant}

\item{log}{Logical; Return log density?}

\item{seed}{if not NULL, then sets the seed.  This is helfpul given that
the Hutchinson probe vectors are randomly sampled, and comparisons have
lower variance using a fixed seed.}

\item{orthogonalize}{Whether to do two-pass Gram-Schmidt re-normalization
(much slower)}
}
\description{
Calculate log-density of a GMRF using Lanczos recursion
}
\examples{
library(RTMB)
library(Matrix)
library(lanczosRTMB)

# Create a precision for a 1D AR1 process
P = bandSparse( n = 100, k = c(-1,1), diagonals = list(rep(0.5,99),rep(0.5,99)) )
Q = (Diagonal(100) - 0.5*t(P) ) \%*\% (Diagonal(100) - 0.5*P)
x = RTMB:::rgmrf0(n = 1, Q )[,1]

# Exact density using RTMB
dgmrf(x, Q = Q, log = TRUE)

# Lanczos-approximated density
lanczos_dgmrf(x, Q = Q, k = 20, log = TRUE)

}
