% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lanczos.R
\name{lanczos_variance}
\alias{lanczos_variance}
\title{Estimate variance using Lanczos method}
\usage{
lanczos_variance(
  Hv,
  v,
  k = c(25, 30),
  min_spectral_ratio = 1e-10,
  orthogonalize = FALSE
)
}
\arguments{
\item{Hv}{function that calculates the product \code{Hv}}

\item{v}{vector to use when calculating variance, either an indicator for a single
parameter, or a gradient evaluated at the MLE for a derived quantity}

\item{k}{can be a vector}

\item{min_spectral_ratio}{is the ratio of minimum to maximum ratio,
where values < minimum are truncated to minimum
(min_spectral_ratio=0 disables truncation)}

\item{orthogonalize}{Whether to do two-pass Gram-Schmidt re-normalization
(much slower)}
}
\description{
Estimate squared standard error (variance) for a parameter or derived quantity
without forming the Hessian matrix directly
}
