% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmrf.R
\name{logdet_icpl}
\alias{logdet_icpl}
\title{Approximate the log-determinant of a GMRF using incomplete Cholesky}
\usage{
logdet_icpl(X)
}
\description{
Approximate the log-determinant of a GMRF using incomplete Cholesky
}
