% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lanczos.R
\name{make_Hv}
\alias{make_Hv}
\title{Make function to calculate H \%*\% v}
\usage{
make_Hv(obj, par = obj$env$last.par.best, tape)
}
\arguments{
\item{obj}{TMB object (output from \code{TMB::MakeADFun})}

\item{par}{parameter vector \code{p} used when evaluating \code{H}}
}
\description{
Given a TMB object, make a function that efficiently calculates
\code{H \%*\% v} without constructing H itself, and instead using \verb{grad_x( grad_x(f) \%** v)}
given function f(x) that returns the negative log-likelihood.  This can
then be used e.g. in Lanczos methods when H is too large to construct explicitly
}
\examples{
u = rnorm(100)
y = rpois(length(u), exp(u))
nll = function(p) -1 * ( sum(dnorm(p$u,log=TRUE)) + sum(dpois(y,exp(p$u),log=TRUE)) )
obj = RTMB::MakeADFun( nll, list(u=u) )
Hv = make_Hv( obj )
# Confirm
v = rnorm( length(obj$par) )
all.equal( Hv(v)[1,], (obj$he()\%*\%v)[,1] )

}
