% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CG.R
\name{CG}
\alias{CG}
\title{Truncated conjugate gradient using Hessian-vector products}
\usage{
CG(
  b,
  Hq,
  x = 0 * b,
  Minv = Diagonal(n = length(b)),
  max.it = length(b),
  e = 1e-10,
  silent = TRUE
)
}
\arguments{
\item{b}{vector to solve for}

\item{Hq}{efficient Hessian-vector product function, e.g., \link{make_Hq}}

\item{x}{initial guess for solution}

\item{Minv}{preconditioner matrix (uses Identity by default)}

\item{max.it}{maximum iterations (often should be less than \code{length(b)})}

\item{e}{error criterion}

\item{silent}{Be silent or print progress?}
}
\description{
Approximate the soution to a linear system \eqn{Hq = b} using truncated conjugate gradient,
without forming the Hessian matrix directly and instead using Hessian-vector products
e.g., using forwards-on-reverse autodiff
Adapted from \code{mcmcsae::CG} under GPL-3 licence to use Hessian-vector products
}
