% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lanczos.R
\name{lanczos}
\alias{lanczos}
\title{Assemble tri-diagonal matrix}
\usage{
lanczos(Hq, q1, k, orthogonalize = FALSE, tol = 1e-12)
}
\arguments{
\item{Hq}{function that calculates the product \code{H \%*\% q}}

\item{q1}{initial vector used for defining a Kyrlov subspace}

\item{k}{dimension for Kyrlov subspace}

\item{orthogonalize}{Whether to do two-pass Gram-Schmidt re-normalization
(much slower)}

\item{tol}{numerical tolerance for stopping algorithm given that no more terms are identifiable}
}
\description{
Assemble tri-diagonal matrix from alpha and beta from Lanczos method
}
\examples{
H = diag(exp(rnorm(5)))
q = rep(1,5)
Hq = function(q) (H \%*\% q)[,1]

L = lanczos(Hq, q, k = 5, ortho = TRUE)
T = tridiag(L$alpha, L$beta)

# Should match H if and only if L$m = nrow(H)
range(L$Q \%*\% T \%*\% t(L$Q) - H)

}
