% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lanczos.R
\name{lanczos_fixedQ}
\alias{lanczos_fixedQ}
\title{Calculate Lanczos approximation using fixed Q}
\usage{
lanczos_fixedQ(Hq, Q, x = attr(Hq, "env")$x0)
}
\arguments{
\item{Hq}{function that calculates the product \code{H \%*\% q} given probe \code{q} and parameters \code{x}}

\item{Q}{matrix of fixed probes}

\item{x}{parameter vector used when calculating the Hessian matrix}
}
\description{
Recalculate Lanczos approximation using fixed Q to speed up gradient calculations
}
