% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lanczos.R
\name{lanczos_logdet}
\alias{lanczos_logdet}
\title{Estimate log-determinant likelihood using Lanczos method}
\usage{
lanczos_logdet(
  Hq,
  k,
  m,
  n,
  seed = NULL,
  orthogonalize = TRUE,
  return_extra = FALSE
)
}
\arguments{
\item{Hq}{function that calculates the product \code{H \%*\% q}}

\item{k}{dimension for Kyrlov subspace}

\item{m}{number of probe-vectors to use for approximating average and standard
deviation of log-determinant}

\item{n}{length of parameters (and necessary probe-vector)}

\item{seed}{if not NULL, then sets the seed.  This is helfpul given that
the Hutchinson probe vectors are randomly sampled, and comparisons have
lower variance using a fixed seed.}

\item{orthogonalize}{Whether to do two-pass Gram-Schmidt re-normalization
(much slower)}

\item{return_extra}{whether to return probes and other internal constructions.}
}
\description{
Estimate log-determinant of inner Hessian matrix with a stochastic approximation
}
\details{
For a model with independent random effects, the variance of stochastic
trace estimation should be zero
}
