% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lanczos.R
\name{lanczos_MakeADFun}
\alias{lanczos_MakeADFun}
\title{Approximate log-marginal likelihood using Lanczos method}
\usage{
lanczos_MakeADFun(
  func,
  parameters,
  map = NULL,
  random,
  profile = NULL,
  k,
  m = 3,
  inner_optimizer = "newton_CG",
  HVP_method = c("reverse-on-reverse", "sparse", "FD-on-reverse"),
  seed = 123,
  make_gr = TRUE,
  pu_update = c("implicit", "FD", "exact", "implicit_FD"),
  silent = TRUE
)
}
\arguments{
\item{func}{Function taking a parameter list (or parameter vector) as input.}

\item{parameters}{Parameter list (or parameter vector) used by \code{func}.}

\item{map}{List defining how to optionally collect and fix parameters, defined the same
as in TMB}

\item{random}{Character vector defining the random effect parameters. See also \code{regexp}.}

\item{profile}{Parameters to profile out of the likelihood (this subset will be appended to \code{random} with Laplace approximation disabled).}

\item{k}{dimension for Kyrlov subspace}

\item{m}{number of probe-vectors to use for approximating average and standard
deviation of log-determinant}

\item{inner_optimizer}{whether to use \link{newton_CG} or a gradient-based low-memory option
specifically "L-BFGS-B" in \link{optim} to optimize the inner problem}

\item{HVP_method}{method for computing Hessian-vector-product,
passed as \code{method} to \code{make_Hq}}

\item{seed}{if not NULL, then sets the seed.  This is helfpul given that
the Hutchinson probe vectors are randomly sampled, and comparisons have
lower variance using a fixed seed.}

\item{make_gr}{whether to make approximated gradient using fixed probes
(slow for large models)}

\item{pu_update}{when make_gr=TRUE, whether to update random effects for proposed fixed effects
using a finite-difference based on the implicit solution and a conjugate gradient solver,
a FD based on taping the inner newton solution for random effects, or
re-optimizing the joint likelihood with respect to random effects.
These updated values are then used within a finite-difference approximation to the
gradient of the Lanczos approximation to the log-determinant.
pu_update="FD" can be very slow for dense inner-Hessians.}

\item{silent}{Disable all tracing information?}
}
\value{
An object (list), where elements include:
\describe{
\item{par}{parameter-vector of fixed effects}
\item{fn}{function that returns the Lanczos-Laplace approximation given fixed effects.}
\item{gr}{a function that returns the approximated gradient of \code{fn} with respect to fixed effects.}
\item{env}{environment of local variables}
\item{Hq}{function that returns the Hessian-vector product (for use in debugging)}
}
}
\description{
Make a function that returns the Lanczos-Laplace approximation for
a user-supplied joint likelihood and designated random effects
}
\details{
The gradient uses a finite-difference applied to a fixed set of probes,
inspired by Dong et al. (2017) and the \code{stochasticLQ} option in GPyTorch.
Exploration suggests that a useful approximation to the gradient of the
log-marginal likelihood with respect to fixed effects can be calculated using
the gradient of the joint likelihood with respect to the fixed effects,
and a finite-difference approximation to the log-determinant.  The latter
approximation is only performs well when recalculating the Lanczos matrix \eqn{Q}.
}
\examples{
# Simulate lognormal-gamma process
set.seed(123)
library(RTMB)
n = 30
n_sum = 3
u = 0 + rnorm(n)
y = rgamma( n, shape = 1/0.5^2, scale = exp(u) * 0.5^2 )

# Fit as GLMM
what = "jnll"
nll = function(p){
  sumexpu = sum(exp(p$u[seq_len(n_sum)]))
  nll1 = dnorm(p$u, mean=p$mu, sd=exp(p$logsd), log=TRUE)
  nll2 = dgamma(y, shape = 1/exp(2*p$logcv), scale = exp(p$u) * exp(2*p$logcv), log=TRUE)
  jnll = -1 * ( sum(nll1) + sum(nll2) )
  return(jnll)
}
parlist =  list(u=u, mu = 0, logsd = 0, logcv = 0)

# Build
obj = lanczos_MakeADFun( nll, parlist, random = "u", k = 10 )
opt = nlminb( obj$par, obj$fn )

# Compare with RTMB
obj2 = MakeADFun( nll, parlist, random = "u", silent = TRUE )
opt2 = nlminb( obj2$par, obj2$fn, obj2$gr )
opt$par - opt2$par

# Fit again using FD gradient for Lanczos method using fixed probe-recursion
opt3 = optim( obj$par, obj$fn, obj$gr, method = "BFGS" )
opt3$par - opt2$par

}
\references{
Dong, K., Eriksson, D., Nickisch, H., Bindel, D., & Wilson, A. G. (2017). Scalable log determinants for Gaussian process kernel learning. Advances in Neural Information Processing Systems, 30.
\url{https://proceedings.neurips.cc/paper/2017/hash/976abf49974d4686f87192efa0513ae0-Abstract.html}
}
