% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lanczos.R
\name{lanczos_MakeADFun}
\alias{lanczos_MakeADFun}
\title{Approximate log-marginal likelihood using Lanczos method (EXPERIMENTAL)}
\usage{
lanczos_MakeADFun(
  func,
  parameters,
  random,
  k,
  profile = NULL,
  m = 3,
  seed = 123
)
}
\arguments{
\item{func}{Function taking a parameter list (or parameter vector) as input.}

\item{parameters}{Parameter list (or parameter vector) used by \code{func}.}

\item{random}{Character vector defining the random effect parameters. See also \code{regexp}.}

\item{k}{dimension for Kyrlov subspace}

\item{profile}{Parameters to profile out of the likelihood (this subset will be appended to \code{random} with Laplace approximation disabled).}

\item{m}{number of probe-vectors to use for approximating average and standard
deviation of log-determinant}

\item{seed}{if not NULL, then sets the seed.  This is helfpul given that
the Hutchinson probe vectors are randomly sampled, and comparisons have
lower variance using a fixed seed.}
}
\value{
An object (list) of class \code{tinyVAST}. Elements include:
\describe{
\item{par}{parameter-vector of fixed effects}
\item{fn}{function that returns the Lanczos-Laplace approximation given fixed effects}
\item{env}{environment of local variables}
\item{Hq}{function that returns the Hessian-vector product (for use in debugging)}
}
}
\description{
Make a function that returns the Lanczos-Laplace approximation for
a user-supplied joint likelihood and designated random effects
}
\examples{
# Simulate lognormal-gamma process
set.seed(123)
library(RTMB)
n = 30
n_sum = 3
u = 0 + rnorm(n)
y = rgamma( n, shape = 1/0.5^2, scale = exp(u) * 0.5^2 )

# Fit as GLMM
what = "jnll"
nll = function(p){
  sumexpu = sum(exp(p$u[seq_len(n_sum)]))
  nll1 = dnorm(p$u, mean=p$mu, sd=exp(p$logsd), log=TRUE)
  nll2 = dgamma(y, shape = 1/exp(2*p$logcv), scale = exp(p$u) * exp(2*p$logcv), log=TRUE)
  jnll = -1 * ( sum(nll1) + sum(nll2) )
  return(jnll)
}

# Build
obj = lanczos_MakeADFun( nll, list(u=u, mu = 0, logsd = 0, logcv = 0), random = "u", k = 10 )
opt = nlminb( obj$par, obj$fn )

# Compare with RTMB
obj2 = MakeADFun( nll, list(u=u, mu = 0, logsd = 0, logcv = 0), random = "u", silent = TRUE )
opt2 = nlminb( obj2$par, obj2$fn, obj2$gr )
opt$par - opt2$par

}
