% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lanczos.R
\name{lanczos_MakeADFun}
\alias{lanczos_MakeADFun}
\title{Approximate log-marginal likelihood using Lanczos method (EXPERIMENTAL)}
\usage{
lanczos_MakeADFun(
  func,
  parameters,
  random,
  k,
  profile = NULL,
  m = 3,
  seed = 123
)
}
\arguments{
\item{func}{Function taking a parameter list (or parameter vector) as input.}

\item{parameters}{Parameter list (or parameter vector) used by \code{func}.}

\item{random}{As \link[TMB]{MakeADFun}.}

\item{k}{dimension for Kyrlov subspace}

\item{profile}{As \link[TMB]{MakeADFun}.}

\item{m}{number of probe-vectors to use for approximating average and standard
deviation of log-determinant}

\item{seed}{if not NULL, then sets the seed.  This is helfpul given that
the Hutchinson probe vectors are randomly sampled, and comparisons have
lower variance using a fixed seed.}
}
\value{
An object (list) of class \code{tinyVAST}. Elements include:
\describe{
\item{par}{parameter-vector of fixed effects}
\item{fn}{function that returns the Lanczos-Laplace approximation given fixed effects}
\item{env}{environment of local variables}
\item{Hv}{function that returns the Hessian-vector product (for use in debugging)}
}
}
\description{
Make a function that returns the Lanczos-Laplace approximation for
a user-supplied joint likelihood and designated random effects
}
