% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lanczos.R
\name{make_Hq}
\alias{make_Hq}
\title{Make function to calculate H \%*\% q}
\usage{
make_Hq(obj, uhat = obj$env$last.par.best, tape)
}
\arguments{
\item{obj}{TMB object (output from \code{TMB::MakeADFun})}

\item{par}{parameter vector \code{u} used when evaluating \code{H}}
}
\description{
Given a TMB object, make a function that efficiently calculates
\code{H \%*\% q} without constructing H itself, and instead using \verb{grad_u( grad_u(f) \%** q)}
given function f(x) that returns the negative log-likelihood given \code{x = u} with fixed \code{v}.  This can
then be used e.g. in Lanczos methods when H is too large to construct explicitly
}
\examples{
u = rnorm(100)
y = rpois(length(u), exp(u))
nll = function(p) -1 * ( sum(dnorm(p$u,log=TRUE)) + sum(dpois(y,exp(p$u),log=TRUE)) )
obj = RTMB::MakeADFun( nll, list(u=u), silent = TRUE )
Hq = make_Hq( obj )
# Confirm
q = rnorm( length(obj$par) )
all.equal( Hq(q)[1,], (obj$he()\%*\%v)[,1] )

}
