% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CG.R
\name{newton_CG}
\alias{newton_CG}
\title{Nonlinear minimizer using line search with approximated Newton solution}
\usage{
newton_CG(
  par,
  fn,
  gr,
  Hq,
  gr_tol = 1e-08,
  e_ratio = 0.01,
  maxit_newton = 100,
  maxit_CG = min(100, length(par)),
  c1 = 0.01,
  beta = 0.5,
  line_steps = 20,
  smartsearch = TRUE,
  smart_x0 = TRUE,
  ustep = 1,
  power = 0.5,
  u0 = 1e-04,
  stop_if_nonPD = smartsearch,
  tol10 = 0.001,
  diagnostics = FALSE,
  silent = FALSE
)
}
\arguments{
\item{par}{initial parameter vector}

\item{fn}{function to evaluate negative log-likelihood}

\item{gr}{function to evaluate gradient of negative log-likelihood}

\item{Hq}{efficient Hessian-vector product function, e.g., \link{make_Hq}}

\item{gr_tol}{early stopping condition for gradient of Newton solver}

\item{e_ratio}{early stopping condition for error of CG, relative to initial gradient}

\item{maxit_newton}{maximum iterations for Newton solver}

\item{maxit_CG}{maximum iterations for CG solution for each Newton iteration}

\item{c1}{stopping condition for line search given CG solution in each Newton iteration, where
\eqn{1>c1>0} seeks to prevent overshoot and resulting zigzagging.}

\item{beta}{updates in line search stepsize alpha when Armijo sufficient decrease condition fails}

\item{line_steps}{number of steps to explore for linear-search given Newton update}

\item{smartsearch}{Turn on adaptive stepsize algorithm for non-convex problems?}

\item{smart_x0}{whether to use previous solution as x0 for next CG iteration}

\item{ustep}{Adaptive stepsize initial guess between 0 and 1.}

\item{power}{Parameter controlling adaptive stepsize.}

\item{u0}{Parameter controlling adaptive stepsize.}

\item{stop_if_nonPD}{whether to stop CG if recusion is not positive definite}

\item{tol10}{Try to exit if last 10 iterations not improved more than this.}

\item{diagnostics}{whether to provide extra diagnostics for each Newton iteration}

\item{silent}{Be silent or print progress?}
}
\description{
Nonlinear minimizer designed to use cheap Hessian-vector products using
\link{make_Hq}.  The minimizer approximates a Newton update using truncated conjugate
gradient, while adapting a regularization parameter designed and using a
line-search for each step.
}
\details{
This minimizer approximates Newton steps \eqn{x_{i+1} = x_{i} - H(x_i)^{-1} g(x_i)},
where \eqn{H(x_i) = \nabla^2 f(x_i)} is the Hessian matrix and
\eqn{g(x_i) = \nabla f(x_i)} is the gradient of the negative log-likelihood
\eqn{f(x_i)}.
It then uses several strategies for numerical and computational efficiency.
\enumerate{
\item It approximates each Newton step \eqn{H(x_i)^{-1} g(x_i)} using a truncated
conjugate gradient algorithm that involves Hessian-vector products \eqn{H(x_i) v}.
It then recursively improves the solution until a desired accuracy is reached,
controlled by \code{gr_tol}, often using many fewer steps than the full dimension
\code{length(x_i)};
\item It approximates each Newton solution \eqn{H(x_i)^{-1} g(x_i)}
without ever constructing or storing the Hessian \eqn{H(x_i)}, and instead computing
Hessian-vector product \eqn{H(x_i) v = \nabla( v^T \nabla f(x_i) )} using
autodifferentiation in RTMB;
\item After approximating each step direction \eqn{H(x_i)^{-1} g(x_i)},
it uses a line-search algorithm while decreasing the step-size if needed
to find a suitable decrease in the objective function, controlled
\code{line_steps}, \code{beta}, and \code{c1};
\item The CG is monitored to identify whether
the Hessian matrix is positive definite (PD), and if it is not PD then the CG may be
terminated, controlled by \code{stop_if_nonPD}.
\item When using \code{smartsearch = TRUE}, if the Hessian is not PD, then
a regularization \code{ustep} is decreased, corresponding to an increase in
\eqn{t=\frac{1}{u}-1} where the Newton step is actually using \eqn{(H + tI)^{-1} g}
where \eqn{g} is the gradient, and \eqn{1>u>0} corresponds to \eqn{t>0}.
This increase or decrease in \code{ustep} (and associated decrease/increase
in \eqn{t}) is copied from \code{\link[TMB:newton]{TMB::newton()}}. If \code{smartsearch = FALSE} then
\eqn{ustep=1} and \eqn{t=0} such that CG uses the Hessian
corresponding to unregularized Newton steps.
This smartsearch behavior controlled by \code{u0}, \code{ustep},
\code{power}, and \code{tol10}, and it corresponds to an adaptive "trust-region".
}
}
\examples{
library(Matrix)
library(RTMB)

# Settings
set.seed(123)
n = 1000
rho = 0.999

# Simulate AR1 process approaching random walk (i.e., ill-conditioned inner problem)
P = bandSparse( n = n, k = c(-1,1), diagonals = list(rep(0.5,n),rep(0.5,n)) )
Q = (Diagonal(n) - rho * t(P)) \%*\% (400 * Diagonal(n)) \%*\% (Diagonal(n) - rho * P)
x = RTMB:::rgmrf0( n= 1, Q = Q )[,1]
y = rpois( n = n, lambda = exp(1 + x) )
which_seen = sample( seq_len(n), size = n/10, replace = FALSE)
y[-which_seen] = NA

nll = function(p){
  -dgmrf(p$x, Q = Q, log = TRUE) - sum(dpois(y, lambda = exp(1 + p$x), log=TRUE), na.rm=TRUE)
}
parlist = list( x=rnorm(n) )

tape = MakeTape(nll, parlist)
gr = tape$jacfun()
Hq = make_Hq( tape, x = unlist(parlist) )
H = gr$jacfun(sparse = TRUE)

start_time = Sys.time()
opt1 = optim(
  par = unlist(parlist),
  fn = tape,
  gr = gr,
  method = "L-BFGS-B",
  control = list(
    maxit = 1e5
  )
)
opt1$runtime = Sys.time() - start_time

opt2 = newton_CG(
  par = unlist(parlist),
  fn = tape,
  gr = gr,
  Hq = Hq,
  maxit_newton = 1e4,
  ustep = 0.01,
  silent = TRUE
)

# Compare the estimates and speed
matplot( cbind(opt1$par, opt2$par), type = "l", col = c("black","blue","red"), lty = "solid" )
c(opt1$runtime, opt2$runtime)
# newton_CG finds a slightly better fit
c(opt1$value, opt2$value)

}
