% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CG.R
\name{newton_CG}
\alias{newton_CG}
\title{Nonlinear minimizer using line search with approximated Newton solution}
\usage{
newton_CG(
  par,
  fn,
  gr,
  Hq,
  gr_tol = 1e-08,
  e_ratio = 0.01,
  maxit_newton = 100,
  maxit_CG = min(100, length(par)),
  c1 = 0.01,
  beta = 0.5,
  silent = FALSE
)
}
\arguments{
\item{par}{initial parameter vector}

\item{fn}{function to evaluate negative log-likelihood}

\item{gr}{function to evaluate gradient of negative log-likelihood}

\item{Hq}{efficient Hessian-vector product function, e.g., \link{make_Hq}}

\item{gr_tol}{early stopping condition for gradient of Newton solver}

\item{e_ratio}{early stopping condition for error of CG, relative to initial gradient}

\item{maxit_newton}{maximum iterations for Newton solver}

\item{maxit_CG}{maximum iterations for CG solution for each Newton iteration}

\item{c1}{stopping condition for line search given CG solution in each Newton iteration}

\item{beta}{updates in line search stepsize alpha when Armijo sufficient decrease condition fails}

\item{silent}{Be silent or print progress?}
}
\description{
Nonlinear minimizer designed for cheap Hessian-vector products using
\link{make_Hq}, involving iterating a linear search along the truncated conjugate gradient
for a Newton.
}
