% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CG.R
\name{newton_CG}
\alias{newton_CG}
\title{Nonlinear minimizer using line search with approximated Newton solution}
\usage{
newton_CG(
  par,
  fn,
  gr,
  Hq,
  gr_tol = 1e-08,
  e_ratio = 0.01,
  maxit_newton = 100,
  maxit_CG = min(100, length(par)),
  c1 = 0.01,
  beta = 0.5,
  silent = FALSE
)
}
\arguments{
\item{par}{initial parameter vector}

\item{fn}{function to evaluate negative log-likelihood}

\item{gr}{function to evaluate gradient of negative log-likelihood}

\item{Hq}{efficient Hessian-vector product function, e.g., \link{make_Hq}}

\item{gr_tol}{early stopping condition for gradient of Newton solver}

\item{e_ratio}{early stopping condition for error of CG, relative to initial gradient}

\item{maxit_newton}{maximum iterations for Newton solver}

\item{maxit_CG}{maximum iterations for CG solution for each Newton iteration}

\item{c1}{stopping condition for line search given CG solution in each Newton iteration}

\item{beta}{updates in line search stepsize alpha when Armijo sufficient decrease condition fails}

\item{silent}{Be silent or print progress?}
}
\description{
Nonlinear minimizer designed for cheap Hessian-vector products using
\link{make_Hq}, involving iterating a linear search along the truncated conjugate gradient
for a Newton.
}
\examples{
library(Matrix)
library(RTMB)

# Settings
n = 10^4
rho = 0.99

# Simulate AR1 process approaching random walk (i.e., ill-conditioned inner problem)
P = bandSparse( n = n, k = c(-1), diagonals = list(rep(1,n)) )
Q = (Diagonal(n) - rho * t(P)) \%*\% (Diagonal(n) - rho * P)
x = RTMB:::rgmrf0( n= 1, Q = Q )[,1]
y = x + 0.1 * rnorm(n)
which_seen = sample( seq_len(n), size = n/10, replace = FALSE)
y[-which_seen] = NA

nll = function(p){
  -dgmrf(p$x, Q = Q, log = TRUE) - sum(dnorm(y, p$x, sd = 0.1, log=TRUE), na.rm=TRUE)
}
parlist = list( x=rnorm(n) )

tape = MakeTape(nll, parlist)
gr = tape$jacfun()
Hq = make_Hq( tape, x = unlist(parlist) )
H = gr$jacfun(sparse = TRUE)

start_time = Sys.time()
opt1 = optim(
  par = unlist(parlist),
  fn = tape,
  gr = gr,
  method = "L-BFGS-B",
  control = list(
    maxit = 1e4
  )
)
opt1$runtime = Sys.time() - start_time

opt2 = newton_CG(
  par = unlist(parlist),
  fn = tape,
  gr = gr,
  Hq = Hq
)

# Compare the estimates and speed
matplot( cbind(opt1$par, opt2$par), type = "l", col = c("black","blue","red"), lty = "solid" )
c(opt1$runtime, opt2$runtime)

}
