% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CG.R
\name{newton_CG}
\alias{newton_CG}
\title{Nonlinear minimizer using line search with approximated Newton solution}
\usage{
newton_CG(
  par,
  fn,
  gr,
  Hq,
  gr_tol = 1e-08,
  e_ratio = 0.01,
  maxit_newton = 100,
  maxit_CG = min(100, length(par)),
  c1 = 0.01,
  beta = 0.5,
  line_steps = 20,
  smartsearch = FALSE,
  ustep = 1,
  power = 0.5,
  u0 = 1e-04,
  stop_if_nonPD = smartsearch,
  tol10 = 0.001,
  diagnostics = FALSE,
  silent = FALSE
)
}
\arguments{
\item{par}{initial parameter vector}

\item{fn}{function to evaluate negative log-likelihood}

\item{gr}{function to evaluate gradient of negative log-likelihood}

\item{Hq}{efficient Hessian-vector product function, e.g., \link{make_Hq}}

\item{gr_tol}{early stopping condition for gradient of Newton solver}

\item{e_ratio}{early stopping condition for error of CG, relative to initial gradient}

\item{maxit_newton}{maximum iterations for Newton solver}

\item{maxit_CG}{maximum iterations for CG solution for each Newton iteration}

\item{c1}{stopping condition for line search given CG solution in each Newton iteration}

\item{beta}{updates in line search stepsize alpha when Armijo sufficient decrease condition fails}

\item{line_steps}{number of steps to explore for linear-search given Newton update}

\item{smartsearch}{Turn on adaptive stepsize algorithm for non-convex problems?}

\item{ustep}{Adaptive stepsize initial guess between 0 and 1.}

\item{power}{Parameter controlling adaptive stepsize.}

\item{u0}{Parameter controlling adaptive stepsize.}

\item{stop_if_nonPD}{whether to stop CG if recusion is not positive definite}

\item{tol10}{Try to exit if last 10 iterations not improved more than this.}

\item{diagnostics}{whether to provide extra diagnostics for each Newton iteration}

\item{silent}{Be silent or print progress?}
}
\description{
Nonlinear minimizer designed to use cheap Hessian-vector products using
\link{make_Hq}.  The minimizer approximates a Newton update using truncated conjugate
gradient, while adapting a regularization parameter designed and using a
line-search for each step.
}
\details{
This minimizer approximates Newton steps, but solving each Newton solution using a
truncated conjugate gradient using Hessian-vector products without ever constructing
the Hessian matrix directly.  It then uses three separate tricks for numerical and
computational efficiency.
\enumerate{
\item For each CG, it recursively improves the solution until a desired accuracy is reached;
\item For each Newton-step, it uses a line-search algorithm, while decreasing the step-size
to find a decrease in objective function.
\item The CG is monitored to identify whether
the Hessian matrix is positive definite (PD), and if it is not PD then the CG may be
terminated.
\item  When using \code{smartsearch = TRUE}, if the Hessian is not PD, then
a regularization \code{ustep} is decreased, corresponding to an increase in \code{t}
where the Newton step is actually using \eqn{(H + tI)^{-1} g} where \eqn{g} is the
gradient.  This increase or decrease in \code{ustep} (and associated decrease/increase
in \eqn{t}) is copied from \code{TMB::newton}. If \code{smartsearch = FALSE} then
\eqn{t=0} and CG uses the Hessian corresponding to unregularized Newton steps.
}
}
\examples{
library(Matrix)
library(RTMB)

# Settings
n = 10^3
rho = 0.99

# Simulate AR1 process approaching random walk (i.e., ill-conditioned inner problem)
P = bandSparse( n = n, k = c(-1), diagonals = list(rep(1,n)) )
Q = (Diagonal(n) - rho * t(P)) \%*\% (Diagonal(n) - rho * P)
x = RTMB:::rgmrf0( n= 1, Q = Q )[,1]
y = rpois( n = n, lambda = exp(1 + x) )
which_seen = sample( seq_len(n), size = n/10, replace = FALSE)
y[-which_seen] = NA

nll = function(p){
  -dgmrf(p$x, Q = Q, log = TRUE) - sum(dpois(y, lambda = exp(1 + p$x), log=TRUE), na.rm=TRUE)
}
parlist = list( x=rnorm(n) )

tape = MakeTape(nll, parlist)
gr = tape$jacfun()
Hq = make_Hq( tape, x = unlist(parlist) )
H = gr$jacfun(sparse = TRUE)

start_time = Sys.time()
opt1 = optim(
  par = unlist(parlist),
  fn = tape,
  gr = gr,
  method = "L-BFGS-B",
  control = list(
    maxit = 1e4
  )
)
opt1$runtime = Sys.time() - start_time

opt2 = newton_CG(
  par = unlist(parlist),
  fn = tape,
  gr = gr,
  Hq = Hq,
  maxit_newton = 1e4,
  gr_tol = 1e-4
)

# Compare the estimates and speed
matplot( cbind(opt1$par, opt2$par), type = "l", col = c("black","blue","red"), lty = "solid" )
c(opt1$runtime, opt2$runtime)

}
