% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lanczos.R
\name{tridiag}
\alias{tridiag}
\title{Assemble tri-diagonal symmetric matrix}
\usage{
tridiag(alpha, beta)
}
\arguments{
\item{alpha}{vector for diagonal}

\item{beta}{vector for off-diagonal}
}
\description{
Assemble tri-diagonal matrix from alpha and beta from Lanczos method
}
